requirements.txt
Django==1.10.4
multi-key-dict==2.0.3
pbr==1.10.0
python-jenkins==0.4.13
six==1.10.0
